The first assignment of Stanford's machine learning course.
Initialised git repository a bit late but I have already completed the first part ex1.m of this first assignment. This involved completing the
computeCost and gradientDecent functions to implement a linear regression algorithm in one variable. Took me a bit of time to work out
the maths and write the program in vectorised form.