Updated the file for feature normalisation.
In the data set the number of bedrooms feature is orders of magnitude smaller than the house sizes feature. As a result gradient descent
could run very slowly as a result. So we normalise the features by subtracting the mean and divinding by standard deviation (of each
feature) for that feature.