In this function I had to use the fmincg advanced optimization function to find the thetas for each classifier (so 10 different thetas in our exercise of classifing digits).
I did this with a loop through 1:num_labels where num_labels is 10 in our case and transposed each of the theta and add it to our matrix of
all_thetas which contains each theta as a row. The idea for one vs all is that we train each classifier seperately and assign a value of 1
to y if it is the correct classification and y = 0 for anything else. For example, for the digit 3 we will only return y=1 if the training
example is actually 3 and y=0 otherwise. Hence, can use y == K in the fmincg optimization function where K = 1:num_labels.