Completed the function dataset3Params which used the validation set to select optimum values of C and sigma.
I used a selection of 8 different values suggested in the instructions. I used a double for loop to go through all the possible C and sigma
in each calling other functions to: first train the model; then return a vector of predictions using the model; finally calculate the error
on the predicted result. Then append the values of C, sigma and error to a matrix. Finally, outside the for loop just had to find the
corresponding C and sigma for the lowest error.