Completed the function findClosestCentroids.
My basic idea was to loop through all the training examples and for each find the distance to each centroid and append to an array
distances, then find the mimnimum distance and assign the corresponding centroid to ind(i) (i goes from 1 to m). But to actually
find the dinstances I made a big mistake here, in which I hard coded it to the 2 dimensional case and it took me ages to realise what was
wrong. Only after I looked at the test cases where X was 10x5 and centroid 4x5 I realised that this is 5 dimensional but I had
only calculated coordinates x1 (with X(i,1)) and x2, c1, c2... Then, I instantly realised that I can just find the coordinate as a vector
for any dimension with X(i,:) and then I can just do a simple vector subtraction and take the inner product to find the squared distance.